+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added options --target-resource-group and --force-move-resource-group to backup ship, restore and schedule enable.
+- Added --layer-list argument to resource-definition clone
 
 ## [1.23.2] - 2024-09-25